Added dependencies and refactored imported packages.
