+++ service/src/main/java/com/epam/rft/atsy/service/passwordchange/validation/PasswordValidator.java
@@ -1,7 +1,7 @@
 package com.epam.rft.atsy.service.passwordchange.validation;
 
 import com.epam.rft.atsy.service.domain.PasswordChangeDTO;
-import com.epam.rft.atsy.web.exception.PasswordValidationException;
+import com.epam.rft.atsy.service.exception.PasswordValidationException;
 
 public interface PasswordValidator {
     boolean validate(PasswordChangeDTO passwordChangeDTO) throws PasswordValidationException;
