+++ service/src/main/java/com/epam/rft/atsy/service/passwordchange/validation/impl/PasswordUniqueRule.java
@@ -3,7 +3,7 @@ package com.epam.rft.atsy.service.passwordchange.validation.impl;
 import com.epam.rft.atsy.service.PasswordChangeService;
 import com.epam.rft.atsy.service.domain.PasswordChangeDTO;
 import com.epam.rft.atsy.service.passwordchange.validation.PasswordValidationRule;
-import com.epam.rft.atsy.web.security.UserDetailsAdapter;
+import com.epam.rft.atsy.service.security.UserDetailsAdapter;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Component;
