+++ web/src/main/java/com/epam/rft/atsy/web/controllers/PasswordChangeController.java
@@ -5,9 +5,9 @@ import com.epam.rft.atsy.service.UserService;
 import com.epam.rft.atsy.service.domain.PasswordChangeDTO;
 import com.epam.rft.atsy.service.domain.PasswordHistoryDTO;
 import com.epam.rft.atsy.service.domain.UserDTO;
-import com.epam.rft.atsy.web.exception.PasswordValidationException;
-import com.epam.rft.atsy.web.passwordchange.validation.PasswordValidator;
-import com.epam.rft.atsy.web.security.UserDetailsAdapter;
+import com.epam.rft.atsy.service.exception.PasswordValidationException;
+import com.epam.rft.atsy.service.passwordchange.validation.PasswordValidator;
+import com.epam.rft.atsy.service.security.UserDetailsAdapter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
